backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
openai>=1.12.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
sqlalchemy==2.0.23
httpx>=0.27.0
matplotlib==3.8.2
seaborn==0.13.0

